完善观测输出与结果文件格式
1. observations1.txt 增加 SNR 输出，并统一小数位格式。
2. 将 SNR 从整数改为 double 保存，避免解码时丢失小数。
3. spp_solution.txt 参考坐标改为与 pos 文件一致，便于 ENU 误差分析。
4. 调整 spp_solution.txt 表头格式，使其更便于与参考文件对比。
5. 为 VS 工程添加 /utf-8 编译选项，避免中文注释导致编码错误。

+++ SPP/obs.h
@@ -9,7 +9,7 @@ struct obsd_t {
     double L[NFREQ + NEXOBS];               // 相位观测（周）
     double P[NFREQ + NEXOBS];               // 伪距观测（米）
     float D[NFREQ + NEXOBS];                // 多普勒/码率
-    unsigned short SNR[NFREQ + NEXOBS];     // 信噪比
+    double SNR[NFREQ + NEXOBS];             // 信噪比（dB-Hz）
     unsigned char LLI[NFREQ + NEXOBS];      // 周跳标记
     unsigned char code[NFREQ + NEXOBS];     // 观测码类型
 } ;
